float

+++ kernels.cu
@@ -1,18 +1,18 @@
 #include "defines.h"
 
 __global__ void fluidMovement(
-	const double* xVel0,
-	const double* yVel0,
-	const double* zVel0,
-	const double* xArea,
-	const double* yArea,
-	const double* zArea,
-	double* mass0,
-	const double* volume,
+	const float* xVel0,
+	const float* yVel0,
+	const float* zVel0,
+	const float* xArea,
+	const float* yArea,
+	const float* zArea,
+	float* mass0,
+	const float* volume,
 	char* warpInfo,
-	double deltaTime,
-	double velFlux,
-	double areaFlux,
+	float deltaTime,
+	float velFlux,
+	float areaFlux,
 	int xThreads,
 	int yThreads,
 	int zThreads,
@@ -48,30 +48,30 @@ __global__ void fluidMovement(
 	                    : (tz == bdz - 1 ? index + sizeBlock * gdx * gdy - bdx * bdy * tz
 	                                     : index + bdx * bdy);
 
-	const double xVelEntry = (x == 0) ? velFlux * areaFlux : xVel0[index] * xArea[index];
-	const double xVelExit = (xIndex_1B == -1) ? velFlux * areaFlux : xVel0[xIndex_1B] * xArea[xIndex_1B];
+	const float xVelEntry = (x == 0) ? velFlux * areaFlux : xVel0[index] * xArea[index];
+	const float xVelExit = (xIndex_1B == -1) ? velFlux * areaFlux : xVel0[xIndex_1B] * xArea[xIndex_1B];
 
-	const double yVelEntry = (y == 0) ? 0.0 : yVel0[index] * yArea[index];
-	const double yVelExit = (yIndex_1B == -1) ? 0.0 : yVel0[yIndex_1B] * yArea[yIndex_1B];
+	const float yVelEntry = (y == 0) ? 0.0 : yVel0[index] * yArea[index];
+	const float yVelExit = (yIndex_1B == -1) ? 0.0 : yVel0[yIndex_1B] * yArea[yIndex_1B];
 
-	const double zVelEntry = (z == 0) ? 0.0 : zVel0[index] * zArea[index];
-	const double zVelExit = (zIndex_1B == -1) ? 0.0 : zVel0[zIndex_1B] * zArea[zIndex_1B];
+	const float zVelEntry = (z == 0) ? 0.0 : zVel0[index] * zArea[index];
+	const float zVelExit = (zIndex_1B == -1) ? 0.0 : zVel0[zIndex_1B] * zArea[zIndex_1B];
 
 	mass0[index] += (xVelEntry - xVelExit + yVelEntry - yVelExit + zVelEntry - zVelExit) * deltaTime;
 }
 
 __global__ void recalculateVelocities(
-	double* xVel0,
-	double* yVel0,
-	double* zVel0,
-	const double* mass0,
-	const double* xArea,
-	const double* yArea,
-	const double* zArea,
-	const double* volume,
-	double beginMass,
-	double deltaTime,
-	double damping,
+	float* xVel0,
+	float* yVel0,
+	float* zVel0,
+	const float* mass0,
+	const float* xArea,
+	const float* yArea,
+	const float* zArea,
+	const float* volume,
+	float beginMass,
+	float deltaTime,
+	float damping,
 	float blocking,
 	int xThreads,
 	int yThreads,
@@ -91,64 +91,64 @@ __global__ void recalculateVelocities(
 
 	const int index = sizeBlock * ((int)blockIdx.x + gdx * ((int)blockIdx.y + (int)blockIdx.z * gdy)) + (tx + bdx * (ty + tz * bdy));
 
-	double v = volume[index];
+	float v = volume[index];
 
 	if (v == 0) return;
-	double newVelX = xVel0[index];
-	double newVelY = yVel0[index];
-	double newVelZ = zVel0[index];
+	float newVelX = xVel0[index];
+	float newVelY = yVel0[index];
+	float newVelZ = zVel0[index];
 
-	const double m = mass0[index];
+	const float m = mass0[index];
 	//const double m = 5;
-	const double rho = m / v;
+	const float rho = m / v;
 
 	/*
 	int T = 300;
 	double R = 8.314;
 	double M = 0.02897;
 	*/
-	constexpr double TR_M = 86095.961; // T*R/M
+	constexpr float TR_M = 86095.961; // T*R/M
 
 	// X ---
-	const double xA = xArea[index];
+	const float xA = xArea[index];
 
 	if (xA != 0 && x != 0)
 	{
 		const int i_xm1 = (tx == 0) ? index - sizeBlock + bdx - 1 : index - 1;
-		const double m_xm1 = mass0[i_xm1];
-		const double v_xm1 = volume[i_xm1];
-		const double deltaP = (m_xm1 / v_xm1 - rho) * TR_M;
-		double ax = deltaP * xA / (m + m_xm1);
+		const float m_xm1 = mass0[i_xm1];
+		const float v_xm1 = volume[i_xm1];
+		const float deltaP = (m_xm1 / v_xm1 - rho) * TR_M;
+		float ax = deltaP * xA / (m + m_xm1);
 		newVelX = (newVelX + ax * deltaTime) * damping;
 		if ((newVelX > 0 && m_xm1 <= 0) || (newVelX < 0 && m <= 0)) newVelX *= blocking;
 	}
 
 
 	// Y ---
-	const double yA = yArea[index];
+	const float yA = yArea[index];
 
 	if (yA != 0 && y != 0)
 	{
 		const int i_ym1 = (ty == 0) ? index - sizeBlock * gdx + bdx * (bdy - 1) : index - bdx;
-		const double m_ym1 = mass0[i_ym1];
-		const double v_ym1 = volume[i_ym1];
-		const double deltaP = (m_ym1 / v_ym1 - rho) * TR_M;
-		double ay = deltaP * yA / (m + m_ym1);
+		const float m_ym1 = mass0[i_ym1];
+		const float v_ym1 = volume[i_ym1];
+		const float deltaP = (m_ym1 / v_ym1 - rho) * TR_M;
+		float ay = deltaP * yA / (m + m_ym1);
 		newVelY = (newVelY + ay * deltaTime) * damping;
 		if ((newVelY > 0 && m_ym1 <= 0) || (newVelY < 0 && m <= 0)) newVelY *= blocking;
 	}
 
 
 	// Z ---
-	const double zA = zArea[index];
+	const float zA = zArea[index];
 
 	if (zA != 0 && z != 0)
 	{
 		const int i_zm1 = (tz == 0) ? index - sizeBlock * gdx * gdy + bdx * bdy * (bdz - 1) : index - bdx * bdy;
-		const double m_zm1 = mass0[i_zm1];
-		const double v_zm1 = volume[i_zm1];
-		const double deltaP = (m_zm1 / v_zm1 - rho) * TR_M;
-		double az = deltaP * zA / (m + m_zm1);
+		const float m_zm1 = mass0[i_zm1];
+		const float v_zm1 = volume[i_zm1];
+		const float deltaP = (m_zm1 / v_zm1 - rho) * TR_M;
+		float az = deltaP * zA / (m + m_zm1);
 		newVelZ = (newVelZ + az * deltaTime) * damping;
 		if ((newVelZ > 0 && m_zm1 <= 0) || (newVelZ < 0 && m <= 0)) newVelZ *= blocking;
 	}
